Module refactor
