+++ riccati/tests/test-riccati.py
@@ -1,6 +1,10 @@
 # Unit tests for riccati.py
 import numpy as np
-import riccati
+from riccati.chebutils import cheb, integrationm, quadwts
+from riccati.solversetup import solversetup
+from riccati.evolve import solve, osc_evolve, nonosc_evolve
+from riccati.step import osc_step, nonosc_step
+from riccati.stepsize import choose_osc_stepsize, choose_nonosc_stepsize
 import scipy.special as sp
 import mpmath
 import warnings
@@ -10,11 +14,11 @@ def test_integration():
     a = 3.0
     f = lambda x: np.sin(a*x + 1.0)
     df = lambda x: a*np.cos(a*x + 1.0)
-    D, x = riccati.cheb(n)
+    D, x = cheb(n)
     dfs = df(x)
     fs = f(x)
     fs -= fs[-1]
-    Im = riccati.integrationm(n+1)
+    Im = integrationm(n+1)
     fs_est = Im @ dfs
     maxerr = max(np.abs((fs_est - fs)/fs))
     assert maxerr < 1e-8
@@ -22,7 +26,7 @@ def test_integration():
 def test_denseoutput():
     w = lambda x: np.sqrt(x)
     g = lambda x: np.zeros_like(x)
-    info = riccati.setup(w, g, n = 32, p = 32)
+    info = solversetup(w, g, n = 32, p = 32)
     xi = 1e0
     xf = 1e6
     eps = 1e-12
@@ -31,7 +35,7 @@ def test_denseoutput():
     dyi = -sp.airy(-xi)[1] - 1j*sp.airy(-xi)[3]
     Neval = int(1e2)
     xeval = np.linspace(xi, xf, Neval) 
-    xs, ys, dys, ss, ps, stypes, yeval = riccati.solve(info, xi, xf, yi, dyi,\
+    xs, ys, dys, ss, ps, stypes, yeval = solve(info, xi, xf, yi, dyi,\
                                                        xeval = xeval,\
                                                        eps = eps, epsh = epsh)
     ytrue = np.array([mpmath.airyai(-x) + 1j*mpmath.airybi(-x) for x in xeval])
@@ -42,7 +46,7 @@ def test_denseoutput():
 def test_denseoutput_xbac():
     w = lambda x: np.sqrt(x)
     g = lambda x: np.zeros_like(x)
-    info = riccati.setup(w, g, n = 32, p = 32)
+    info = solversetup(w, g, n = 32, p = 32)
     xi = 1e0
     xf = 1e6
     eps = 1e-12
@@ -51,7 +55,7 @@ def test_denseoutput_xbac():
     dyi = -sp.airy(-xi)[1] - 1j*sp.airy(-xi)[3]
     Neval = int(1e2)
     xeval = np.linspace(xf, xi, Neval) 
-    xs, ys, dys, ss, ps, stypes, yeval = riccati.solve(info, xi, xf, yi, dyi,\
+    xs, ys, dys, ss, ps, stypes, yeval = solve(info, xi, xf, yi, dyi,\
                                                        xeval = xeval,\
                                                        eps = eps, epsh = epsh)
     ytrue = np.array([mpmath.airyai(-x) + 1j*mpmath.airybi(-x) for x in xeval])
@@ -62,7 +66,7 @@ def test_denseoutput_xbac():
 def test_denseoutput_warn():
     w = lambda x: np.sqrt(x)
     g = lambda x: np.zeros_like(x)
-    info = riccati.setup(w, g, n = 32, p = 32)
+    info = solversetup(w, g, n = 32, p = 32)
     xi = 1e0
     xf = 1e6
     eps = 1e-12
@@ -74,7 +78,7 @@ def test_denseoutput_warn():
     # Turn on warnings
     warnings.simplefilter("always")
     with warnings.catch_warnings(record = True) as w:
-        xs, ys, dys, ss, ps, stypes, yeval = riccati.solve(info, xi, xf, yi, dyi,\
+        xs, ys, dys, ss, ps, stypes, yeval = solve(info, xi, xf, yi, dyi,\
                                                            xeval = xeval,\
                                                            eps = eps, epsh = epsh)
         assert "outside the integration range" in str(w[0].message) 
@@ -84,8 +88,8 @@ def test_quadwts():
     f = lambda x: np.sin(a*x + 1.0)
     df = lambda x: a*np.cos(a*x + 1.0)
     for i in [16, 32]:
-        D, x = riccati.cheb(i)
-        w = riccati.quadwts(i)
+        D, x = cheb(i)
+        w = quadwts(i)
         maxerr = w.dot(df(x)) - (f(1) - f(-1)) 
         assert  maxerr < 1e-8
 
@@ -94,14 +98,14 @@ def test_cheb():
     f = lambda x: np.sin(a*x + 1.0)
     df = lambda x: a*np.cos(a*x + 1.0)
     for i in [16, 32]:
-        D, x = riccati.cheb(i)
+        D, x = cheb(i)
         maxerr = max(np.abs(np.matmul(D, f(x)) - df(x))) 
         assert  maxerr < 1e-8
 
 def test_osc_step():
     w = lambda x: np.sqrt(x) 
     g = lambda x: np.zeros_like(x)
-    info = riccati.setup(w, g)
+    info = solversetup(w, g)
     x0 = 10.0
     h = 20.0
     epsres = 1e-12
@@ -112,20 +116,20 @@ def test_osc_step():
     dy0 = -sp.airy(-x0)[1]
     y_ana = sp.airy(-(x0+h))[0]
     dy_ana = -sp.airy(-(x0+h))[1]
-    y, dy, res, success, phase = riccati.osc_step(info, x0, h, y0, dy0, epsres = epsres)
+    y, dy, res, success, phase = osc_step(info, x0, h, y0, dy0, epsres = epsres)
     y_err = np.abs((y - y_ana)/y_ana)
     assert y_err < 1e-8 and res < epsres
 
 def test_nonosc_step():
     w = lambda x: np.sqrt(x) 
     g = lambda x: np.zeros_like(x)
-    info = riccati.setup(w, g)
+    info = solversetup(w, g)
     x0 = 1.0
     h = 0.5
     eps = 1e-12
     y0 = sp.airy(-x0)[2]
     dy0 = - sp.airy(-x0)[3]
-    y, dy, err, success = riccati.nonosc_step(info, x0, h, y0, dy0, epsres = eps)
+    y, dy, err, success = nonosc_step(info, x0, h, y0, dy0, epsres = eps)
     y_ana = sp.airy(-x0-h)[2]
     dy_ana = -sp.airy(-x0-h)[3]
     y_err = np.abs((y - y_ana)/y_ana)
@@ -135,14 +139,14 @@ def test_nonosc_step():
 def test_solve_airy():
     w = lambda x: np.sqrt(x)
     g = lambda x: np.zeros_like(x)
-    info = riccati.setup(w, g, n = 32, p = 32)
+    info = solversetup(w, g, n = 32, p = 32)
     xi = 1e0
     xf = 1e6
     eps = 1e-12
     epsh = 1e-13
     yi = sp.airy(-xi)[0] + 1j*sp.airy(-xi)[2]
     dyi = -sp.airy(-xi)[1] - 1j*sp.airy(-xi)[3]
-    xs, ys, dys, ss, ps, stypes = riccati.solve(info, xi, xf, yi, dyi, eps = eps, epsh = epsh)
+    xs, ys, dys, ss, ps, stypes = solve(info, xi, xf, yi, dyi, eps = eps, epsh = epsh)
     xs = np.array(xs)
     ys = np.array(ys)
     ytrue = np.array([mpmath.airyai(-x) + 1j*mpmath.airybi(-x) for x in xs])
@@ -154,14 +158,14 @@ def test_solve_airy():
 def test_solve_airy_backwards():
     w = lambda x: np.sqrt(x)
     g = lambda x: np.zeros_like(x)
-    info = riccati.setup(w, g, n = 32, p = 32)
+    info = solversetup(w, g, n = 32, p = 32)
     xi = 1e6
     xf = 1e0
     eps = 1e-12
     epsh = 1e-13
     yi = sp.airy(-xi)[0] + 1j*sp.airy(-xi)[2]
     dyi = -sp.airy(-xi)[1] - 1j*sp.airy(-xi)[3]
-    xs, ys, dys, ss, ps, stypes = riccati.solve(info, xi, xf, yi, dyi,\
+    xs, ys, dys, ss, ps, stypes = solve(info, xi, xf, yi, dyi,\
                                                 eps = eps, epsh = epsh,\
                                                 hard_stop = True)
     xs = np.array(xs)
@@ -175,7 +179,7 @@ def test_solve_airy_backwards():
 def test_denseoutput_backwards_xfor():
     w = lambda x: np.sqrt(x)
     g = lambda x: np.zeros_like(x)
-    info = riccati.setup(w, g, n = 32, p = 32)
+    info = solversetup(w, g, n = 32, p = 32)
     xi = 1e6
     xf = 1e0
     eps = 1e-12
@@ -184,7 +188,7 @@ def test_denseoutput_backwards_xfor():
     dyi = -sp.airy(-xi)[1] - 1j*sp.airy(-xi)[3]
     Neval = int(1e2)
     xeval = np.linspace(xi, xf, Neval) 
-    xs, ys, dys, ss, ps, stypes, yeval = riccati.solve(info, xi, xf, yi, dyi,\
+    xs, ys, dys, ss, ps, stypes, yeval = solve(info, xi, xf, yi, dyi,\
                                                        xeval = xeval,\
                                                        eps = eps, epsh = epsh,\
                                                        hard_stop = True)
@@ -197,7 +201,7 @@ def test_denseoutput_backwards_xfor():
 def test_denseoutput_backwards_xback():
     w = lambda x: np.sqrt(x)
     g = lambda x: np.zeros_like(x)
-    info = riccati.setup(w, g, n = 32, p = 32)
+    info = solversetup(w, g, n = 32, p = 32)
     xi = 1e6
     xf = 1e0
     eps = 1e-12
@@ -206,7 +210,7 @@ def test_denseoutput_backwards_xback():
     dyi = -sp.airy(-xi)[1] - 1j*sp.airy(-xi)[3]
     Neval = int(1e2)
     xeval = np.linspace(xf, xi, Neval) 
-    xs, ys, dys, ss, ps, stypes, yeval = riccati.solve(info, xi, xf, yi, dyi,\
+    xs, ys, dys, ss, ps, stypes, yeval = solve(info, xi, xf, yi, dyi,\
                                                        xeval = xeval,\
                                                        eps = eps, epsh = epsh,\
                                                        hard_stop = True)
@@ -229,8 +233,8 @@ def test_solve_burst():
     dyi = burstdy(xi)
     eps = 1e-10
     epsh = 1e-12
-    info = riccati.setup(w, g, n = 32, p = 32)
-    xs, ys, dys, ss, ps, types = riccati.solve(info, xi, xf, yi, dyi, eps = eps, epsh = epsh)
+    info = solversetup(w, g, n = 32, p = 32)
+    xs, ys, dys, ss, ps, types = solve(info, xi, xf, yi, dyi, eps = eps, epsh = epsh)
     xs = np.array(xs)
     ys = np.array(ys)
     ytrue = bursty(xs)
@@ -241,7 +245,7 @@ def test_solve_burst():
 def test_osc_evolve():
     w = lambda x: np.sqrt(x)
     g = lambda x: np.zeros_like(x)
-    info = riccati.setup(w, g, n = 32, p = 32)
+    info = solversetup(w, g, n = 32, p = 32)
     xi = 1e2
     xf = 1e6
     eps = 1e-12
@@ -254,12 +258,12 @@ def test_osc_evolve():
     info.y = np.array([yi, dyi])
     # Always necessary for setting info.wn, info.gn, and for getting an initial stepsize
     hi = 2*xi
-    hi = riccati.choose_osc_stepsize(info, xi, hi, epsh = epsh)
+    hi = choose_osc_stepsize(info, xi, hi, epsh = epsh)
     info.h = hi
     # Not necessary here because info.x is already xi, but in general it might be:
     info.x = xi
     while info.x < xf:
-        status = riccati.osc_evolve(info, info.x, xf, info.h, info.y, epsres = eps, epsh = epsh)
+        status = osc_evolve(info, info.x, xf, info.h, info.y, epsres = eps, epsh = epsh)
         if status != 1:
             break
         xs.append(info.x)
@@ -274,7 +278,7 @@ def test_osc_evolve():
 def test_nonosc_evolve():
     w = lambda x: np.sqrt(x)
     g = lambda x: np.zeros_like(x)
-    info = riccati.setup(w, g, n = 32, p = 32)
+    info = solversetup(w, g, n = 32, p = 32)
     xi = 1e0
     xf = 4e1
     eps = 1e-12
@@ -287,12 +291,12 @@ def test_nonosc_evolve():
     info.y = np.array([yi, dyi])
     # Necessary for getting an initial stepsize
     hi = 1/w(xi)
-    hi = riccati.choose_nonosc_stepsize(info, xi, hi, epsh = epsh)
+    hi = choose_nonosc_stepsize(info, xi, hi, epsh = epsh)
     info.h = hi
     # Not necessary here because info.x is already xi, but in general it might be:
     info.x = xi
     while info.x < xf:
-        status = riccati.nonosc_evolve(info, info.x, xf, info.h, info.y, epsres = eps, epsh = epsh)
+        status = nonosc_evolve(info, info.x, xf, info.h, info.y, epsres = eps, epsh = epsh)
         if status != 1:
             break
         xs.append(info.x)
